Streamline test fixtures and vlidation helpers
