+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "p2pool-v2"
+name = "p2poolv2"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.83"
