+++ src/lib.rs
@@ -18,4 +18,6 @@ pub mod command;
 pub mod config;
 pub mod node;
 pub mod shares;
+
+#[cfg(test)]
 pub mod test_utils;
