+++ src/shares/chain/actor.rs
@@ -301,8 +301,8 @@ mock! {
 mod tests {
     use super::*;
     use crate::shares::miner_message::Gbt;
-    use crate::test_utils::fixtures::random_hex_string;
-    use crate::test_utils::fixtures::simple_miner_share;
+    use crate::test_utils::random_hex_string;
+    use crate::test_utils::simple_miner_share;
     use rust_decimal_macros::dec;
     use tempfile::tempdir;
